fixup! fix(react-tag-picker): remove default icon from headless state
