+++ change/@fluentui-react-tag-picker-73122b43-f401-4b38-ba87-5f92358110f5.json
@@ -1,6 +1,6 @@
 {
   "type": "patch",
-  "comment": "fix: separate headless TagPicker state from styled icon defaults",
+  "comment": "fix: omit the default TagPicker expand icon from headless state",
   "packageName": "@fluentui/react-tag-picker",
   "email": "vgenaev@gmail.com",
   "dependentChangeType": "patch"
